// webapp/src/app/Model/MedicalEquipment.ts
export class MedicalEquipment {


    createdBy: String;
    type: String;
    typeofEquipment: String;
    medicalEquipmentName: String;
    city: String;
    avalabilityPlace: String;
    address: String;
    contactPersonName: String;
    contactMobileNumber: String;
    isVerified: String;
    verifiedBy: String;

}